+++ src/main/java/plugily/projects/villagedefense/arena/ArenaEvents.java
@@ -34,6 +34,7 @@ import org.bukkit.event.player.PlayerRespawnEvent;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.potion.PotionEffectType;
 import org.bukkit.scheduler.BukkitRunnable;
+import pl.plajerlair.commonsbox.minecraft.compat.ServerVersion;
 import plugily.projects.villagedefense.ConfigPreferences;
 import plugily.projects.villagedefense.Main;
 import plugily.projects.villagedefense.api.StatsStorage;
@@ -45,7 +46,6 @@ import plugily.projects.villagedefense.handlers.items.SpecialItem;
 import plugily.projects.villagedefense.handlers.language.Messages;
 import plugily.projects.villagedefense.handlers.reward.Reward;
 import plugily.projects.villagedefense.user.User;
-import plugily.projects.villagedefense.utils.ServerVersion.Version;
 import plugily.projects.villagedefense.utils.Utils;
 
 import java.util.ArrayList;
@@ -62,7 +62,7 @@ public class ArenaEvents implements Listener {
 
   public ArenaEvents(Main plugin) {
     this.plugin = plugin;
-    if (Version.isCurrentEqualOrLower(Version.v1_12_R1)) {
+    if (ServerVersion.Version.isCurrentEqualOrLower(ServerVersion.Version.v1_12_R1)) {
       plugin.getServer().getPluginManager().registerEvents(new Events1_8(), plugin);
     } else {
       plugin.getServer().getPluginManager().registerEvents(new Events1_9(), plugin);
@@ -265,12 +265,12 @@ public class ArenaEvents implements Listener {
     new BukkitRunnable() {
       @Override
       public void run() {
-        if (Version.isCurrentEqual(Version.v1_11_R1) || arena.getArenaState() == ArenaState.ENDING) {
+        if (ServerVersion.Version.isCurrentEqual(ServerVersion.Version.v1_11_R1) || arena.getArenaState() == ArenaState.ENDING) {
           this.cancel();
           return;
         }
         if (user.isSpectator()) {
-          if (Version.isCurrentEqualOrHigher(Version.v1_16_R1)) {
+          if (ServerVersion.Version.isCurrentEqualOrHigher(ServerVersion.Version.v1_16_R1)) {
             user.getPlayer().sendActionBar(TextComponent.fromLegacyText(plugin.getChatManager().colorMessage(Messages.DIED_RESPAWN_IN_NEXT_WAVE)));
           } else {
             user.getPlayer().spigot().sendMessage(ChatMessageType.ACTION_BAR, TextComponent.fromLegacyText(plugin.getChatManager().colorMessage(Messages.DIED_RESPAWN_IN_NEXT_WAVE)));