+++ src/main/java/plugily/projects/villagedefense/handlers/sign/SignManager.java
@@ -19,7 +19,6 @@
 package plugily.projects.villagedefense.handlers.sign;
 
 import org.apache.commons.lang.StringUtils;
-
 import org.bukkit.ChatColor;
 import org.bukkit.Location;
 import org.bukkit.block.Block;
@@ -32,6 +31,7 @@ import org.bukkit.event.block.BlockBreakEvent;
 import org.bukkit.event.block.SignChangeEvent;
 import org.bukkit.event.player.PlayerInteractEvent;
 import org.jetbrains.annotations.Nullable;
+import pl.plajerlair.commonsbox.minecraft.compat.ServerVersion;
 import pl.plajerlair.commonsbox.minecraft.compat.XMaterial;
 import pl.plajerlair.commonsbox.minecraft.configuration.ConfigUtils;
 import pl.plajerlair.commonsbox.minecraft.serialization.LocationSerializer;
@@ -43,7 +43,6 @@ import plugily.projects.villagedefense.arena.ArenaState;
 import plugily.projects.villagedefense.handlers.language.LanguageManager;
 import plugily.projects.villagedefense.handlers.language.Messages;
 import plugily.projects.villagedefense.utils.Debugger;
-import plugily.projects.villagedefense.utils.ServerVersion.Version;
 import plugily.projects.villagedefense.utils.constants.Constants;
 
 import java.util.ArrayList;
@@ -209,31 +208,31 @@ public class SignManager implements Listener {
           switch (arenaSign.getArena().getArenaState()) {
             case WAITING_FOR_PLAYERS:
               behind.setType(XMaterial.WHITE_STAINED_GLASS.parseMaterial());
-              if (Version.isCurrentLower(Version.v1_13_R1)) {
+              if (ServerVersion.Version.isCurrentLower(ServerVersion.Version.v1_13_R1)) {
                 Block.class.getMethod("setData", byte.class).invoke(behind, (byte) 0);
               }
               break;
             case STARTING:
               behind.setType(XMaterial.YELLOW_STAINED_GLASS.parseMaterial());
-              if (Version.isCurrentLower(Version.v1_13_R1)) {
+              if (ServerVersion.Version.isCurrentLower(ServerVersion.Version.v1_13_R1)) {
                 Block.class.getMethod("setData", byte.class).invoke(behind, (byte) 4);
               }
               break;
             case IN_GAME:
               behind.setType(XMaterial.ORANGE_STAINED_GLASS.parseMaterial());
-              if (Version.isCurrentLower(Version.v1_13_R1)) {
+              if (ServerVersion.Version.isCurrentLower(ServerVersion.Version.v1_13_R1)) {
                 Block.class.getMethod("setData", byte.class).invoke(behind, (byte) 1);
               }
               break;
             case ENDING:
               behind.setType(XMaterial.GRAY_STAINED_GLASS.parseMaterial());
-              if (Version.isCurrentLower(Version.v1_13_R1)) {
+              if (ServerVersion.Version.isCurrentLower(ServerVersion.Version.v1_13_R1)) {
                 Block.class.getMethod("setData", byte.class).invoke(behind, (byte) 7);
               }
               break;
             case RESTARTING:
               behind.setType(XMaterial.BLACK_STAINED_GLASS.parseMaterial());
-              if (Version.isCurrentLower(Version.v1_13_R1)) {
+              if (ServerVersion.Version.isCurrentLower(ServerVersion.Version.v1_13_R1)) {
                 Block.class.getMethod("setData", byte.class).invoke(behind, (byte) 15);
               }
               break;