+++ src/main/java/plugily/projects/villagedefense/utils/NMS.java
@@ -3,9 +3,8 @@ package plugily.projects.villagedefense.utils;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.java.JavaPlugin;
-
+import pl.plajerlair.commonsbox.minecraft.compat.ServerVersion;
 import plugily.projects.villagedefense.Main;
-import plugily.projects.villagedefense.utils.ServerVersion.Version;
 
 @SuppressWarnings("deprecation")
 public abstract class NMS {
@@ -13,7 +12,7 @@ public abstract class NMS {
 	private static final Main PLUGIN = JavaPlugin.getPlugin(Main.class);
 
 	public static void hidePlayer(Player to, Player p) {
-		if (Version.isCurrentEqualOrHigher(Version.v1_13_R1)) {
+		if (ServerVersion.Version.isCurrentEqualOrHigher(ServerVersion.Version.v1_13_R1)) {
 			to.hidePlayer(PLUGIN, p);
 		} else {
 			to.hidePlayer(p);
@@ -21,7 +20,7 @@ public abstract class NMS {
 	}
 
 	public static void showPlayer(Player to, Player p) {
-		if (Version.isCurrentEqualOrHigher(Version.v1_13_R1)) {
+		if (ServerVersion.Version.isCurrentEqualOrHigher(ServerVersion.Version.v1_13_R1)) {
 			to.showPlayer(PLUGIN, p);
 		} else {
 			to.showPlayer(p);